sincronizaciones

+++ app/Http/Controllers/InventarioController.php
@@ -58,28 +58,81 @@ class InventarioController extends Controller
             $nombrenuevo = $value['inventario']['name'];
 
             $producto = Inventario::select('id')->where('inventory_id', $inventory_id)->orderBy('id', 'desc')->first();
-            $idunventario = $producto['id'];
-            $invdeletes = Inventario::where('inventory_id', $inventory_id)->where('id', "!=", $idunventario)->get();
-            foreach ($invdeletes as $value) {
-              $preciodelete = Precio::where('inventario_id', $value['id'])->first();
-              if (isset($preciodelete['id'])) {
-                $preciodelete->delete();
+
+            if (isset($producto->id)) {
+              $idunventario = $producto['id'];
+              $invdeletes = Inventario::where('inventory_id', $inventory_id)->where('id', "!=", $idunventario)->get();
+              foreach ($invdeletes as $value) {
+                $preciodelete = Precio::where('inventario_id', $value['id'])->first();
+                if (isset($preciodelete['id'])) {
+                  $preciodelete->delete();
+                }
               }
-            }
-            $producto = Inventario::where('inventory_id', $inventory_id)->where('id', '!=', $idunventario)->forceDelete();
+              $producto = Inventario::where('inventory_id', $inventory_id)->where('id', '!=', $idunventario)->forceDelete();
 
-            $inventarioexist = Inventario_piso_venta::where('piso_venta_id', $request->idpisoventa)->where('inventario_id', $idunventario)->first();
-            //return $inventarioexist;
-            if (isset($inventarioexist->id)) {
-              $inventarioexist->cantidad = $cantidadnueva;
-              $inventarioexist->save();
-
-              $productoupdate = Inventario::where('inventory_id', $inventory_id)->update(['name' => $nombrenuevo]);
-
+              $inventarioexist = Inventario_piso_venta::where('piso_venta_id', $request->idpisoventa)->where('inventario_id', $idunventario)->first();
               //return $inventarioexist;
-              //return "inventarioexist";
-              //return $idunventario;
-              //$inventario->inventario->name = $nombrenuevo;
+              if (isset($inventarioexist->id)) {
+                $inventarioexist->cantidad = $cantidadnueva;
+                $inventarioexist->save();
+
+                $productoupdate = Inventario::where('inventory_id', $inventory_id)->update(['name' => $nombrenuevo]);
+
+              } else {
+                $inventario = new Inventario_piso_venta();
+                $inventario->inventario_id = $idunventario;
+                $inventario->piso_venta_id = $request->idpisoventa;
+                $inventario->cantidad = $cantidadnueva;
+                $inventario->save();
+              }
+            } else {
+
+              $inventarioval = Inventory::select('id')->where('id', $value['inventario']['inventory_id'])->first();
+
+              if (isset($inventarioval->id)) {
+                $articulo = new Inventario();
+                $articulo->name = $value['inventario']['name'];
+                $articulo->unit_type_mayor = $value['inventario']['unit_type_mayor'];
+                $articulo->unit_type_menor = $value['inventario']['unit_type_menor'];
+                $articulo->inventory_id = $value['inventario']['inventory_id'];
+                $articulo->status = $value['inventario']['status'];
+                $articulo->piso_venta_id = $value['inventario']['piso_venta_id'];
+                $articulo->save();
+
+                $inventariopv = Inventario_piso_venta::select('id')->where('inventario_id', $articulo->id)->first();
+
+                if (isset($inventariopv->id)) {
+                  $inventariopv->cantidad = $cantidadnueva;
+                  $inventariopv->save();
+                } else {
+                  $inventario = new Inventario_piso_venta();
+                  $inventario->inventario_id = $articulo->id;
+                  $inventario->piso_venta_id = $request->idpisoventa;
+                  $inventario->cantidad = $cantidadnueva;
+                  $inventario->save();
+                }
+
+                $precioval = Product::where('inventory_id', $value['inventario']['inventory_id'])->first();
+
+                //return response()->json($precioval);
+
+                if (isset($precioval->id)) {
+                  //REGISTRAMOS LOS PRECIOS
+                  $precio = new Precio();
+                  $precio->costo = $precioval['cost'];
+                  $precio->iva_porc = $precioval['iva_percent'];
+                  $precio->iva_menor = $precioval['retail_iva_amount'];
+                  $precio->sub_total_menor = $precioval['retail_total_price'] - $precioval['retail_iva_amount'];
+                  $precio->total_menor = $precioval['retail_total_price'];
+                  $precio->iva_mayor = $precioval['wholesale_iva_amount'];
+                  $precio->sub_total_mayor = $precioval['wholesale_packet_price'];
+                  $precio->total_mayor = $precio->sub_total_mayor + $precio->iva_mayor;
+                  $precio->oferta = $precioval['oferta'];
+                  $precio->inventario_id = $articulo->id;
+                  $precio->save();
+                }
+
+              }
 
             }
 
@@ -115,14 +168,45 @@ class InventarioController extends Controller
 
     public function ultimo_inventory()
     {
-
-        $inventory = Inventory::select('id')->orderBy('id', 'desc')->first();
+      // BK DE SINRONIZAION
+        /*$inventory = Inventory::select('id')->orderBy('id', 'desc')->first();
 
         if (isset($inventory->id)) {
           return response()->json($inventory->id);
         } else {
           return '0';
+        }*/
+      // BK DE SINRONIZAION
+      try {
+
+        DB::beginTransaction();
+
+        $inventoryCreated = Inventory::select('created_at')->orderBy('created_at', 'desc')->first();
+        $inventoryUpdated = Inventory::select('updated_at')->orderBy('updated_at', 'desc')->first();
+        $productUpdated = Product::select('updated_at')->orderBy('updated_at', 'desc')->first();
+        $inventoryDeleted = Inventory::onlyTrashed()->select('deleted_at')->orderBy('deleted_at', 'desc')->first();
+
+        if (isset($inventoryCreated->created_at)) {
+          $created = date('Y-m-d H:i:s', strtotime($inventoryCreated->created_at));
+          $updated = date('Y-m-d H:i:s', strtotime($inventoryUpdated->updated_at));
+          $updatedProd = date('Y-m-d H:i:s', strtotime($productUpdated->updated_at));
+          if (isset($inventoryDeleted->deleted_at)) {
+            $deleted = date('Y-m-d H:i:s', strtotime($inventoryDeleted->deleted_at));
+          } else {
+            $deleted = 0;
+          }
+          DB::commit();
+          return response()->json(['created' => $created, 'updated' => $updated, 'deleted' => $deleted, 'updatedProd' => $updatedProd]);
+        } else {
+          DB::commit();
+          return '0';
         }
+
+      } catch (Exception $e) {
+        DB::rollback();
+        return response()->json($e);
+      }
+
 
 
     }
@@ -137,7 +221,6 @@ class InventarioController extends Controller
 
     public function store_inventory(Request $request)
     {
-      //return response()->json($request->productos);
         try{
 
             DB::beginTransaction();
@@ -153,39 +236,16 @@ class InventarioController extends Controller
                 $inventory->qty_per_unit = $producto['qty_per_unit'];
                 $inventory->status = $producto['status'];
                 $inventory->total_qty_prod = $producto['total_qty_prod'];
+                $inventory->created_at = $producto['created_at'];
                 $inventory->save();
 
 
 
                 if ($producto['product'] != null) {
 
-                  /*$dolar = Dolar::orderby('id','DESC')->first();
-                  $datadolar = $dolar['price'];
-
-
-                  $cost = $producto['product']['cost']*$datadolar;
-                  //= $producto['product']['iva_percent'];
-                  //= $producto['product']['retail_margin_gain'];
-                  $retail_total_price = $producto['product']['retail_total_price']*$datadolar;
-                  //= $producto['product']['retail_iva_amount'];
-                  //= $producto['product']['image'];
-                  //= $producto['product']['wholesale_margin_gain'];
-                  $wholesale_packet_price = round($producto['product']['wholesale_packet_price'], 2)*$datadolar;
-                  //return response()->json($wholesale_packet_price);
-                  $wholesale_total_individual_price = $producto['product']['wholesale_total_individual_price']*$datadolar;
-                  $wholesale_total_packet_price = round($producto['product']['wholesale_total_packet_price'], 2)*$datadolar;
-                  //$arrayErrors = array($cost, $retail_total_price, $wholesale_packet_price, $wholesale_total_individual_price, $wholesale_total_packet_price, $datadolar);
-                  //= $producto['product']['wholesale_iva_amount'];
-                  //= $producto['product']['oferta'];
-                  //= $producto['product']['inventory_id'];*/
-
-                  /*$cost = round($cost, 2);
-                  $retail_total_price = round($retail_total_price, 2);
-                  $wholesale_packet_price = round($wholesale_packet_price, 2);
-                  $wholesale_total_individual_price = round($wholesale_total_individual_price, 2);
-                  $wholesale_total_packet_price = round($wholesale_total_packet_price, 2);*/
-
                   try {
+
+                    //REGISTRAMOS LOS PRECIOS PRODUCT
 
                     $product = new Product();
                     $product->cost = $producto['product']['cost'];
@@ -203,33 +263,14 @@ class InventarioController extends Controller
                     $product->inventory_id = $producto['product']['inventory_id'];
                     $product->save();
 
-                    //REGISTRAMOS LOS PRECIOS PRODUCT
-                    /*$product = new Product();
-                    $product->cost = $cost;
-                    $product->iva_percent = $producto['product']['iva_percent'];
-                    $product->retail_margin_gain = $producto['product']['retail_margin_gain'];
-                    $product->retail_total_price = $retail_total_price;
-                    $product->retail_iva_amount = $producto['product']['retail_iva_amount'];
-                    $product->image = $producto['product']['image'];
-                    $product->wholesale_margin_gain = $producto['product']['wholesale_margin_gain'];
-                    $product->wholesale_packet_price = $wholesale_packet_price;
-                    $product->wholesale_total_individual_price = $wholesale_total_individual_price;
-                    $product->wholesale_total_packet_price = $wholesale_total_packet_price;
-                    $product->wholesale_iva_amount = $producto['product']['wholesale_iva_amount'];
-                    $product->oferta = $producto['product']['oferta'];
-                    $product->inventory_id = $producto['product']['inventory_id'];
-                    $product->save();*/
-
-
                   } catch (Exception $e) {
                     DB::rollback();
                     return response()->json($e);
                   }
 
-
                 }
             }
-            //DB::commit();
+
             DB::commit();
 
             return response()->json(true);
@@ -240,6 +281,124 @@ class InventarioController extends Controller
             return response()->json($e);
         }
 
+    }
+
+    public function update_inventory(Request $request)
+    {
+        try{
+
+            DB::beginTransaction();
+            foreach ($request->productos as $producto) {
+
+              $inventory = Inventory::find($producto['id']);
+
+                $inventory->product_name = $producto['product_name'];
+                $inventory->description = $producto['description'];
+                $inventory->quantity = $producto['quantity'];
+                $inventory->unit_type = $producto['unit_type'];
+                $inventory->unit_type_menor = $producto['unit_type_menor'];
+                $inventory->qty_per_unit = $producto['qty_per_unit'];
+                $inventory->status = $producto['status'];
+                $inventory->total_qty_prod = $producto['total_qty_prod'];
+                $inventory->updated_at = $producto['updated_at'];
+                $inventory->save();
+
+                DB::commit();
+
+            }
+
+            return response()->json("exito");
+
+        }catch(Exception $e){
+
+            DB::rollback();
+            return response()->json($e);
+        }
+
+    }
+
+    public function update_products(Request $request)
+    {
+        try{
+
+            DB::beginTransaction();
+            foreach ($request->precios as $producto) {
+
+              //ACTUALIZAMOS LOS PRECIOS PRODUCT
+
+              $product = Product::where('inventory_id', $producto['inventory_id'])->first();
+
+              if (isset($product->id)) {
+                $product->cost = $producto['cost'];
+                $product->iva_percent = $producto['iva_percent'];
+                $product->retail_margin_gain = $producto['retail_margin_gain'];
+                $product->retail_total_price = $producto['retail_total_price'];
+                $product->retail_iva_amount = $producto['retail_iva_amount'];
+                $product->image = $producto['image'];
+                $product->wholesale_margin_gain = $producto['wholesale_margin_gain'];
+                $product->wholesale_packet_price = $producto['wholesale_packet_price'];
+                $product->wholesale_total_individual_price = $producto['wholesale_total_individual_price'];
+                $product->wholesale_total_packet_price = $producto['wholesale_total_packet_price'];
+                $product->wholesale_iva_amount = $producto['wholesale_iva_amount'];
+                $product->oferta = $producto['oferta'];
+                $product->save();
+              } else {
+                $product = new Product();
+                $product->cost = $producto['cost'];
+                $product->iva_percent = $producto['iva_percent'];
+                $product->retail_margin_gain = $producto['retail_margin_gain'];
+                $product->retail_total_price = $producto['retail_total_price'];
+                $product->retail_iva_amount = $producto['retail_iva_amount'];
+                $product->image = $producto['image'];
+                $product->wholesale_margin_gain = $producto['wholesale_margin_gain'];
+                $product->wholesale_packet_price = $producto['wholesale_packet_price'];
+                $product->wholesale_total_individual_price = $producto['wholesale_total_individual_price'];
+                $product->wholesale_total_packet_price = $producto['wholesale_total_packet_price'];
+                $product->wholesale_iva_amount = $producto['wholesale_iva_amount'];
+                $product->oferta = $producto['oferta'];
+                $product->inventory_id = $producto['inventory_id'];
+                $product->save();
+              }
+
+              DB::commit();
+
+            }
+
+            return response()->json("exito");
+
+        }catch(Exception $e){
+
+            DB::rollback();
+            return response()->json($e);
+        }
+
+    }
+
+    public function deleted_inventory(Request $request)
+    {
+        try{
+
+            DB::beginTransaction();
+            foreach ($request->productos as $producto) {
+
+              //ELIMINAMOS LOS PRODUCT
+
+              $inventory = Inventory::find($producto['id']);
+              if (isset($inventory->id)) {
+                $inventory->delete();
+              }
+
+              DB::commit();
+
+            }
+
+            return response()->json("exito");
+
+        }catch(Exception $e){
+
+            DB::rollback();
+            return response()->json($e);
+        }
 
     }
